+++ R/vapour_vrt.R
@@ -145,4 +145,19 @@ vapour_vrt <- function(x, extent = NULL, projection = NULL,  sds = 1L, bands = N
   out
 }
 
-  
+
+#' Build vrt, special case "-separate"
+#'
+#' @param dsn one or more raster sources
+#'
+#' @return a character string of the builtvrt, multiple sources treated as bands
+#' @export
+#'
+#' @examples
+#' f <- system.file("extdata/sst.tif", package = "vapour", mustWork = TRUE)
+#' vrt <- buildvrt(c(f, vapour_vrt(f)))
+#' writeLines(vrt)
+buildvrt <- function(dsn) {
+   options = c("-separate")
+   raster_buildvrt_cpp(dsn, options)
+ }